workflow: allow manual testing
We periodically have reports of failures in the GitHub Action
where there has been no recent change in Tailscale infrastructure
and no indication of a problem in metrics or monitoring.

Right now we have to generate a pull request to get this test
workflow to run. Allow it to be run manually whenever desired.

Signed-off-by: Denton Gentry <dgentry@tailscale.com>