action.yml: properly remove temporary files after downloading Remove `.tgz` and `.msi` files after they have been downloaded. This was the behaviour of the action previous to `v3.2.0`, but our logic for caching removed the deletion of these files.
Can look at downloading these to temp folders to avoid the more complex
logic here as a follow up, but this should unblock / fix the immediate
regression.

Also, on Windows, write `tailscale.log` to the temp dir.

Also, on MacOS, delete the `tailscale` directory that was used to build the `tailscale(d)` commands.

Updates: #170

Co-authored-by: Mario Minardi <mario@tailscale.com>
Signed-off-by: Percy Wegmann <percy@tailscale.com>