src: check against platform specific versions for latest and unstable
Check against platform and package specific version numbers returned
from pkgs.tailscale.com instead of defaulting to "Version" as this can
be incorrect for certain platforms when a release has only been built
for a subset of platforms.

For example. 1.90.2 is available for linux at the time of writing so
"Version" is 1.90.2, but only 1.90.1 is available for windows which
causes an error when using latest as the version specifier on windows
machines.

Fixes https://github.com/tailscale/github-action/issues/219

Signed-off-by: Mario Minardi <mario@tailscale.com>